refactoring main screen
