+++ src/screens/Main/MainCategories/index.css
